Fixed bug. finite diffs for salamander data ok.
Problem was in objfun where meow was calculated. It worked for T=1 and T>2 but not T=2. Now it has been corrected, and when I check the finite differences in the objfun for the salamander data, it works. This check is not included in the package because the calculation takes too long (several minutes for m=10^6)