Changes and additions for t in R, normal stuff
I've coded the t pdf into R (to be used for tests). Also a little function to calculate the t constant in the pdf. Added the constants to the normal distribution stuff, both for R and C. Some small changes in design doc related to T in R and normal in both R and C. About to move onto creating t in C.