Made a bunch of changes, but haven't found bug
Made a bunch of tidying of help pages
Made a bunch of changes to C code to satisfy
    clang -Weverything -pedantic
    scan-build # the clang static analyzer
But still

    R CMD check glmm_*.tar.gz --run-donttest

crashes and

    cd devel
    R CMD BATCH --vanilla logLik.R

gives a completely mysterious error.
The R "Registering native routines" code says expecting 30 arguments
for C function valgrad, but nothing I can see says 30.
The function itself has 31 arguments, the prototype in the header
file has the same 31 arguments, at three places in glmm_init.c
this function is said to have 31 arguments and nowhere is this
or any other function said to have 30 arguments.

So this looks like memory corruption.  The crash that mystifies
valgrind suggests the same.  We are in trouble.